Wed, Jul 6, 2022, 2:34 AM +00:00

+++ pset4/filter-more/helpers.c
@@ -8,10 +8,10 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
     {
         for (int j = 0; j < width; j++)
         {
-            int avg = round((image[i][j].rgbtRed + image[i][j].rgbtGreen + image[i][j].rgbtBlue) / 3);
-            image[i][j].rgbtRed = avg;
-            image[i][j].rgbtGreen = avg;
-            image[i][j].rgbtBlue = avg;
+            float avg = round((image[i][j].rgbtRed + image[i][j].rgbtGreen + image[i][j].rgbtBlue) / 3);
+            image[i][j].rgbtRed = (int) avg;
+            image[i][j].rgbtGreen = (int) avg;
+            image[i][j].rgbtBlue = (int) avg;
         }
     }
     return;
